backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
slack-sdk==3.27.1
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1